-- ADDED DATABASE SETUP OF FOOD SCREEN -- FETCHING DATA BASED ON RECENT PRODUCT -- ADDED UPSERT OPERATION -- COMPLETED DELETE OPERATION

+++ FoodGrid/app/src/main/java/com/example/foodgrid/model/User.java
@@ -45,7 +45,7 @@ public class User {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
